+++ app/views.py
@@ -1,11 +1,8 @@
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
 from django.template import Context
+from django.db.models import Max
 
 from .models import Blog
-
-def index(request):
-    context = {}
-    return render(request, 'app/index.html', context)
 
 def about(request):
     context = {}
@@ -13,17 +10,27 @@ def about(request):
 
 def blog(request):
     context = {
-        "blogs": Blog.objects.all()[:5],
-        "popular_blog_title": Blog.objects.get(pk=1).title
+        "blogs": Blog.objects.all(),
+        "highest_rated": Blog.objects.all().order_by('-rating').first()
     }
     return render(request, 'app/blog.html', context)
 
-def blog_entry(request):
-    context = {}
+def blog_entry(request, slug):
+    entry = get_object_or_404(Blog, slug=slug)
+    context = {
+        "entry": get_object_or_404(Blog, slug=slug),
+    }
+
+    #   Increment rating on page view
+    entry.rating +=1
+    #   Write updated rating to DB
+    entry.save()
+
     return render(request, 'app/blog_entry.html', context)
 
 def archive(request):
     context = {
-        "blogList": Blog.objects.all()
+        #   Return ALL blog entries ordered by rating descending
+        "blogs": Blog.objects.all().order_by('-rating')
     }
     return render(request, 'app/archive.html', context)