+++ CHANGES.md
@@ -10,8 +10,20 @@
 *   Redefined `models.Blog` schema
 *   Created `fixtures` folder with `blog.json` mock data
 *   Added `.travis.yml` database config
-
-
+*   Implemented `/sitemap.xml`
+*   Added `django.contrib.sitemaps` app for sitemap functionality
+*   Updated mock database fixture data, minor corrections and new entry
+*   Added `face256.png` asset
+*   Modified `base.css` to use face image, hid `#ad` div (for now)
+*   Reverted `blogadamzerella/settings.py` BAZ_SQL_PASS back to `root`
+*   Modified `views.py` contexts to now use `get_object_or_404`
+*   Removed `index.html` temporarily from project
+*   Added rating score increase on page view
+*   Modified archive page order based on blog rating score
+*   Added `</hr>` and most popular blog to `blog.html` under top 5
+*   Modified face redirect now goes to `blog.html` page (default)
+*   Added `Max` package to `views.py`
+*   `Sidebar #ad` is now transparent
 ## 0.0.1
 *   Created `index.html` page with relevant meta info.
 *   Created `Blog` and `Author` models, applied migrations accordingly.